refactor: reanme document_chunk_strategy to chunk_strategy
