+++ src/ingestion_pipeline/orchestration/pipeline.py
@@ -2,7 +2,7 @@
 
 import logging
 
-from ingestion_pipeline.chunking.document_chunk_strategy import ChunkError, ChunkStrategy
+from ingestion_pipeline.chunking.chunk_strategy import ChunkError, ChunkStrategy
 from ingestion_pipeline.chunking.schemas import DocumentMetadata
 from ingestion_pipeline.config import settings
 from ingestion_pipeline.embedding.embedding_generator import EmbeddingError, EmbeddingGenerator
@@ -40,7 +40,7 @@ class Pipeline:
 
         Args:
             textract_processor: Processor to extract data using Textract.
-            chunker: Document chunker with configured strategies.
+            chunker: ChunkStrategy with configured strategies.
             embedding_generator: Generator for creating embeddings from text.
             chunk_indexer: Indexer to store documents in OpenSearch.
             page_indexer: Indexer to store document pages in OpenSearch.
